requirements.txt
praw==7.7.1
pandas==2.2.0
networkx==3.2.1
scikit-learn==1.4.0
vaderSentiment==3.3.2
node2vec==0.4.6
umap-learn==0.5.5
matplotlib==3.8.2
numpy==1.26.4
zstandard==0.22.0
python-qbittorrent==0.4.3
